+++ src/components/menu/Notification.js
@@ -11,16 +11,38 @@ import Popper from '@material-ui/core/Popper';
 import Avatar from '@material-ui/core/Avatar';
 import Divider from "@material-ui/core/Divider";
 import Typography from "@material-ui/core/Typography";
+import Badge from '@material-ui/core/Badge';
+import NotificationsIcon from '@material-ui/icons/Notifications';
+import IconButton from "@material-ui/core/IconButton";
 
 
-export default function Notification({openNotification, notifications,setOpenNotification,handleCloseNotification}){
+export default function Notification(){
 
-
+    const [openNotification, setOpenNotification] = React.useState(null);
+    var notifications = [{img: cancel, text:"El match ha sido cancelado"}
+                        ,{img: timer,text:"Falta 1h para tu que inicie tu match"},
+                        {img: aceppt, text:"Alguien ha aceptado tu reto"}];
+    const handleClickNotification = event => {
+      if (openNotification && openNotification.contains(event.target)) {
+        setOpenNotification(null);
+      } else {
+        setOpenNotification(event.currentTarget);
+      }
+    };
+    
+    const handleCloseNotification = () => {
+      setOpenNotification(null);
+    };
 
 return (
     <div>
-<Popper  open={Boolean(openNotification)} anchorEl={openNotification} transition
-          disablePortal>
+        <IconButton aria-label="show 11 new notifications" color="inherit" onClick={handleClickNotification}>
+            <Badge badgeContent={notifications.length} color="secondary">
+              <NotificationsIcon />
+            </Badge>
+          </IconButton>
+        <Popper  open={Boolean(openNotification)} anchorEl={openNotification} transition
+              >
             {({ TransitionProps, placement }) => (
                     <Grow
                         {...TransitionProps}
